requirements.txt
anyio==3.7.1
bidict==0.22.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.70.1
h11==0.14.0
idna==3.4
netifaces==0.10.6
pydantic==1.10.12
python-engineio==4.7.1
python-socketio==5.7.2
simple-websocket==0.10.1
sniffio==1.3.0
starlette==0.16.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
websockets==10.4
wsproto==1.2.0
